backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-rest-auth==2.2.3
Django==4.0.1
django-debug-toolbar==3.2.4
django-hashid-field==3.3.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
hashids==1.3.1
importlib-metadata==4.10.1
Markdown==3.3.6
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
zipp==3.7.0
